django-backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.42.0
anyio==4.3.0
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.149
botocore==1.34.149
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
dataclasses-json==0.6.4
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-storages==1.14.4
djangorestframework==3.14.0
drf-spectacular==0.27.1
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.12.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.6
inflection==0.5.1
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-text-splitters==0.0.1
langsmith==0.1.38
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
orjson==3.10.0
packaging==23.2
pathlib==1.0.1
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tokenizers==0.21.0
tqdm==4.66.2
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
youtube-transcript-api==0.6.2
